requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
docutils==0.14
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
psycopg2-binary==2.7.7
python-dateutil==1.5
pytz==2018.3
requests==2.20.1
requests-cache==0.4.13
selenium==3.141.0
urllib3==1.24.1
dj-database-url==0.5.0
django-heroku==0.3.1
whitenoise==4.1.2
django-social-share==1.3.2
psycopg2==2.8.2
djangorestframework==3.9.2